refactor(mcp): colocate ai-agent-chat tools with agents and threads
Fold conversation writes into threads.ts and user preferences into
agents.ts so the domain matches evaluations.ts. Share AiAgentScopeArgs
and the ai-agent test harness instead of per-file copies.
