+++ AGENTS.md
@@ -154,7 +154,7 @@ Additional specialized skills are documented in `CLAUDE.md`.
 - Build before ESLint (`dist/` resolves); prefer root `pnpm exec vitest run <files>` over per-package `test:fast`.
 - Changie: `kindFormat: '### {{.Kind}}'`; fragments use kind keys (`feat`/`fix`), not Keep-a-Changelog labels. After `changie merge`, Trunk/Prettier reformats CHANGELOG.md (`*` → `-`, blank lines).
 - `pnpm -r version` needs a clean tree — bump `packages/*/package.json` versions directly if dirty.
-- Agent surface: MCP profiles own mounts via `tools: ToolModule[]` imports (ADR-0022); irrecoverable tool ids stay on `IRRECOVERABLE_TOOL_DENYLIST` (ADR-0004).
+- Agent surface: MCP profiles own mounts via `tools: ToolModule[]` imports (ADR-0022); irrecoverable tool ids stay on `IRRECOVERABLE_TOOL_DENYLIST` (ADR-0004). AI-agent tools colocate by client resource (`agents.ts` inventory+preferences, `threads.ts` read+write, `evaluations.ts`) — do not add `chat.ts`/`preferences.ts`.
 - MCP: `registerToolSafe()` only; CLI: `wrapAction()`; serving profile via `bindServerProfile` + profile `tools` array.
 - MCP progress (SDK v2): emit via `ctx.mcpReq.notify` when `mcpReq._meta.progressToken` is set — there is no top-level `sendNotification` on `ServerContext` (ADR-0023).
 - Guardrails return `_lightdashBlocked`; upstream failures are coded `UPSTREAM_*` / `RATE_LIMITED`, not blocked markers.
